+++ resources/views/livewire/admin/laboratory-component.blade.php
@@ -63,13 +63,16 @@
                                         @endif
                                         <td>{{ date('d-m-Y', strtotime($laboratory->created_at)) }}</td>
                                         <td class="table-action">
-                                            <a href="javascript: void(0);" class="action-ico btn btn-outline-info mx-1"> <i
-                                                    class="bi bi-pencil-square" data-bs-toggle="modal"
+                                            <a href="javascript: void(0);" class="action-ico btn btn-outline-info mx-1">
+                                                <i class="bi bi-pencil-square" data-bs-toggle="modal"
                                                     wire:click="editdata({{ $laboratory->id }})"
                                                     data-bs-target="#editlaboratory"></i></a>
-                                            <a href="javascript: void(0);"
-                                                wire:click="deleteConfirmation({{ $laboratory->id }})"
-                                                class="action-ico btn btn-outline-danger mx-1"> <i class="bi bi-trash"></i></a>
+                                            @if (Auth::user()->hasPermission(['master-access']))
+                                                <a href="javascript: void(0);"
+                                                    wire:click="deleteConfirmation({{ $laboratory->id }})"
+                                                    class="action-ico btn btn-outline-danger mx-1"> <i
+                                                        class="bi bi-trash"></i></a>
+                                            @endif
                                         </td>
                                     </tr>
                                 @endforeach
@@ -113,7 +116,8 @@
                                 </div>
                                 <div class="mb-3">
                                     <label for="shortcod" class="form-label">Short Code</label>
-                                    <input type="text" id="shortcod" class="form-control" wire:model.lazy="short_code">
+                                    <input type="text" id="shortcod" class="form-control"
+                                        wire:model.lazy="short_code">
                                     @error('short_code')
                                         <div class="text-danger text-small">{{ $message }}</div>
                                     @enderror
